reorg blade and segment classes
