+++ opensg/mesh/blade.py
@@ -1,3 +1,5 @@
+from dolfinx.io import gmshio
+from mpi4py import MPI
 import numpy as np
 import dolfinx
 import basix
@@ -7,11 +9,12 @@ from dolfinx.fem import form, petsc, Function, locate_dofs_topological, apply_li
 from scipy.sparse import csr_matrix
 import scipy
 
-from opensg.mesh.segment import SegmentMesh
+import opensg
+from opensg.core import shell as core
 from ..utils import shell as utils
 
 
-class BladeMesh:
+class ShellBladeMesh:
     """A class representing the complete mesh of a wind turbine blade.
     
     This class processes and stores information about a wind turbine blade's mesh,
@@ -228,7 +231,7 @@ class BladeMesh:
         file.close()
         
         # initialize segmentmesh object
-        segment_mesh = SegmentMesh(
+        segment_mesh = BladeSegmentMesh(
             segment_node_labels=segment_node_labels,
             segment_element_labels=segment_element_labels,
             segment_element_layer_id=segment_element_layer_id,
@@ -239,174 +242,671 @@ class BladeMesh:
         return segment_mesh
 
 
-"""
+class SolidBladeSegmentMesh():
+    def __init__(
+        self,
+       # segment_node_labels,
+       # segment_element_labels,
+      #  segment_element_layer_id,
+        segment_index,
+        parent_blade_mesh,
+        msh_file):
+        """This class manages the data and methods for the mesh of a segment of a blade.
+
+        A segment is defined as the part of the blade between two fixed points along the blade span.
+        Given a set of N span points along the blade, there are N-1 segments defined between each consecutive pair
+        of span points. For example, the segment indexed by 0 is defined between the span points indexed by 0 and 1
+
+        Parameters
+        ----------
+        segment_node_labels : array[int]
+            _description_
+        segment_element_labels : array[int]
+            _description_
+        segment_element_layer_id : array[int]
+            _description_
+        segment_index : int
+            Index of the segment of blade
+        parent_blade_mesh : BladeMesh
+            BladeMesh object that SegmentMesh derives from.
+        msh_file : str or Path
+            Path to mesh file to load data from
+        """
+
+     #   self.segment_node_labels = segment_node_labels
+      #  self.segment_element_labels = segment_element_labels
+      #  self.segment_element_layer_id = segment_element_layer_id
+        self.segment_index = segment_index
+        self.blade_mesh = parent_blade_mesh
+        self.elLayID=parent_blade_mesh.elLayID
+        self.material_database=parent_blade_mesh.material_database
+        gmsh.initialize()
+        gmsh.option.setNumber("General.Terminal", 0)    # mesh read output will not be printed
+        self.mesh, self.subdomains, self.boundaries = gmshio.read_from_msh(msh_file, MPI.COMM_WORLD,0, gdim=3)
+        self.original_cell_index = self.mesh.topology.original_cell_index # Original cell Index from mesh file
+   #     lnn = self.subdomains.values[:]-1
+     #   self._generate_layup_id()
+
+        lnn=[]
+        for k in self.original_cell_index:
+             lnn.append(self.elLayID[k])
+
+        self.num_cells = self.mesh.topology.index_map(self.mesh.topology.dim).size_local
+        cells = np.arange(self.num_cells, dtype=np.int32)
+
+        self.subdomains = dolfinx.mesh.meshtags(self.mesh, self.mesh.topology.dim, cells, np.array(lnn,dtype=np.int32))
+
+        # Update elLayID to contain the mapped segment layup IDs (same as StandaloneSolidSegmentMesh)
+        # self.elLayID = np.array(lnn, dtype=np.float64)
+
+        self.tdim = self.mesh.topology.dim
+        self.fdim = self.tdim - 1
+
+      #  self._generate_layup_data()
+   #     self._generate_material_database()
+        self._build_local_orientations()
+        self._build_boundary_submeshes()
+
+        return
+
+    def _build_local_orientations(self):
+        # Local Orientation (DG0 function) of quad mesh element (from yaml data)
+        VV = dolfinx.fem.functionspace(
+            self.mesh, basix.ufl.element(
+            "DG", self.mesh.topology.cell_name(),
+            0, shape=(3, )))
+        EE1, EE2, EE3 = dolfinx.fem.Function(VV), dolfinx.fem.Function(VV), dolfinx.fem.Function(VV)
+
+        orientations = []
+        for i, eo in enumerate(self.blade_mesh.element_orientations):
+                o = []
+                for k in range(9):
+                    o.append(eo[k])
+                orientations.append(o)
+
+        # Store orientation for each element
+        # TODO: clarify variable names. why N?
+        for k, ii in enumerate(self.original_cell_index):
+        # Storing data to DG0 functions 
+            EE2.x.array[3*k], EE2.x.array[3*k+1], EE2.x.array[3*k+2] = orientations[ii][5], orientations[ii][3], orientations[ii][4]  # e2
+            EE3.x.array[3*k], EE3.x.array[3*k+1], EE3.x.array[3*k+2] = orientations[ii][8], orientations[ii][6], orientations[ii][7]   #  e3 
+            EE1.x.array[3*k], EE1.x.array[3*k+1], EE1.x.array[3*k+2] = orientations[ii][2], orientations[ii][0], orientations[ii][1]  # e1    outward normal 
+        self.EE1 = EE1
+        self.EE3 = EE3
+        self.EE2 = EE2
+        frame = [EE1,EE2,EE3]
+        self.frame = frame
+        return frame
+
+    def _build_boundary_submeshes(self):
+        pp = self.mesh.geometry.x
+        x_min,x_max=min(pp[:,0]), max(pp[:,0])
+        mean=0.5*(x_min+x_max)  # Mid origin for taper segments
+
+        blade_length=x_max-x_min
+
+        self.left_origin,self.right_origin,self.taper_origin=[],[],[]
+        self.left_origin.append(float(x_min)/blade_length),self.right_origin.append(float(x_max)/blade_length),self.taper_origin.append(float(mean)/blade_length)
+       # print(float(mean))
+    #    pp[:,0]=pp[:,0]-mean
+
+        is_left_boundary, is_right_boundary = opensg.utils.solid.generate_boundary_markers(
+            min(pp[:,0]), max(pp[:,0]))
+
+        left_facets = dolfinx.mesh.locate_entities_boundary(
+            self.mesh, dim=self.fdim, marker=is_left_boundary)
+        right_facets = dolfinx.mesh.locate_entities_boundary(
+            self.mesh, dim=self.fdim, marker=is_right_boundary)
+
+        left_mesh, left_entity_map, left_vertex_map, left_geom_map = dolfinx.mesh.create_submesh(
+            self.mesh, self.fdim, left_facets)
+        right_mesh, right_entity_map, right_vertex_map, right_geom_map = dolfinx.mesh.create_submesh(
+            self.mesh, self.fdim, right_facets)
+
+        self.left_submesh = {
+            "mesh": left_mesh,
+            "entity_map": left_entity_map,
+            "vertex_map": left_vertex_map,
+            "geom_map": left_geom_map,
+            "marker": is_left_boundary}
+            # "facets": left_facets}
+
+        self.right_submesh = {
+            "mesh": right_mesh,
+            "entity_map": right_entity_map,
+            "vertex_map": right_vertex_map,
+            "geom_map": right_geom_map,
+            "marker": is_right_boundary}
+            # "facets": right_facets}
+
+        self.mesh.topology.create_connectivity(3,2)  # (quad mesh topology, boundary(1D) mesh topology)
+        cell_of_facet_mesh = self.mesh.topology.connectivity(3,2)
+
+        # NOTE: found a different way to determine connectivity that didn't need this block -klb
+        # Cell to Edge connectivity
+        # cell_edge_map = []
+        # for i in range(self.num_cells):
+        #     c = []
+        #     for k in range(4): # 4 is used as number of edges in a quad element
+        #         c.append((cell_of_facet_mesh.array[4*i+k])) 
+        #     cell_edge_map.append(c)
+        # cell_edge_map = np.ndarray.flatten(np.array(cell_edge_map))
+
+        # generate subdomains
+        def _build_boundary_subdomains(boundary_meshdata):
+            boundary_mesh = boundary_meshdata["mesh"]
+            boundary_entity_map = boundary_meshdata["entity_map"]
+            boundary_marker = boundary_meshdata["marker"]
+            boundary_VV = dolfinx.fem.functionspace(
+                boundary_mesh, basix.ufl.element("DG", boundary_mesh.topology.cell_name(), 0, shape=(3, )))
+
+            boundary_e1 = dolfinx.fem.Function(boundary_VV)
+            boundary_e2 = dolfinx.fem.Function(boundary_VV)
+            boundary_e3 = dolfinx.fem.Function(boundary_VV)
+
+            boundary_facets = dolfinx.mesh.locate_entities(boundary_mesh, self.fdim, boundary_marker)
+
+            # TODO: review the subdomain assingments with akshat
+            boundary_subdomains = []
+            el_facets=6
+            for i, xx in enumerate(boundary_entity_map):
+                # assign subdomain
+                # 4 is for number of nodes in quad element
+                # NOTE: we should find a different way to do this that doesn't assume quad elements if
+                #    we plan to expand to other elements. -klb
+                idx = int(np.where(cell_of_facet_mesh.array==xx)[0]/el_facets)
+                boundary_subdomains.append(self.subdomains.values[idx])
+                # assign orientation
+                for j in range(3):
+                    boundary_e1.x.array[3*i+j] = self.EE1.x.array[3*idx+j]
+                    boundary_e2.x.array[3*i+j] = self.EE2.x.array[3*idx+j]
+                    boundary_e3.x.array[3*i+j] = self.EE3.x.array[3*idx+j]
+
+            boundary_frame = [boundary_e1, boundary_e2, boundary_e3]
+            boundary_subdomains = np.array(boundary_subdomains, dtype=np.int32)
+            boundary_num_cells = boundary_mesh.topology.index_map(boundary_mesh.topology.dim).size_local
+            boundary_cells = np.arange(boundary_num_cells, dtype=np.int32)
+            boundary_subdomains = dolfinx.mesh.meshtags(boundary_mesh, boundary_mesh.topology.dim, boundary_cells, boundary_subdomains)
+
+            return boundary_subdomains, boundary_frame, boundary_facets
+            # Mapping the orinetation data from quad mesh to boundary. The alternative is to use local_frame_1D(self.left_submesh["mesh"]).
+            # Either of both can be used in local_boun subroutine 
+
+        self.left_submesh["subdomains"], self.left_submesh["frame"], self.left_submesh["facets"] = _build_boundary_subdomains(self.left_submesh)
+        self.right_submesh["subdomains"], self.right_submesh["frame"], self.right_submesh["facets"] = _build_boundary_subdomains(self.right_submesh)
+
+       # def compute_boundary_stiffness_timo(self):
+           #     left_stiffness=opensg.compute_solidtimo_boun(
+           #         self.material_database[0],
+           #         self.left_submesh)[1]
+
+          #      right_stiffness=opensg.compute_solidtimo_boun(
+          #          self.material_database[0],
+          #          self.right_submesh)[1]    
+
+      #          return  left_stiffness, right_stiffness
+      #  self.left_stiffness, self.right_stiffness=compute_boundary_stiffness_timo(self)
+        self.meshdata = {
+            "mesh": self.mesh,
+            "subdomains": self.subdomains,
+            "frame": self.frame,
+            }
+
+        return
+
+
+class SolidBladeMesh:
+    """This class processes and stores information about a wind turbine blade's mesh
+    """
+    def __init__(self, mesh_data):
+        """
+
+        Parameters
+        ----------
+        mesh_data : dict
+            dictionary of mesh data loaded in from the output of pynumad mesher
+        """
+        self._mesh_data = mesh_data
+
+        self.nodes = mesh_data['nodes']
+        self.num_nodes = len(self.nodes)
+
+        self.elements = mesh_data['elements']
+        self.num_elements = len(self.elements)
+
+        self.sets = mesh_data["sets"]
+        self.materials = mesh_data["materials"]
+   #     self.sections = mesh_data["sections"]
+        self.element_orientations = mesh_data["elementOrientations"]
+        self._generate_layup_id()
+        self._generate_material_database()
+
+    def _generate_layup_id(self):
+        layCt=-1
+        self.mat_name=[]
+        self.elLayID=np.zeros((self.num_elements))
+        for es in self.sets['element']:
+            if es['labels'][0] is not None:
+                self.mat_name.append(es['name'])
+                layCt += 1
+                for eli in es['labels']:
+                    self.elLayID[eli-1]=layCt
+        return
+
+    def _generate_material_database(self):
+   #     material_database = dict()
+        material_parameters,density=[],[]
+        mat_names=[material['name'] for material in self.materials]
+
+        for i, mat in enumerate(self.mat_name):
+            es=self.materials[mat_names.index(mat)]
+            material_parameters.append(np.array((np.array(es['E']),np.array(es['G']),es['nu'])).flatten())
+            density.append(es['rho'])
+          #  material_dict = dict()
+
+          #  material_dict["id"] = i
+
+           # elastic = material['elastic']
+           # material_dict["E"] = elastic['E']
+           # material_dict["G"] = elastic['G']
+           # material_dict["nu"] = elastic['nu']
+           # material_dict["rho"] = elastic['rho']
+         #   material_database[material['name']] = material_dict
+        #######
+        ############            
+        self.material_database = material_parameters,density
+        return
+
+    def generate_segment_mesh(self, segment_index, filename):
+
+        file = open(filename,'w')
+
+        file.write('$MeshFormat\n2.2 0 8\n$EndMeshFormat\n$Nodes\n')
+       # newNumNds = np.max(segment_node_labels)
+
+        file.write(str(self.num_nodes) + '\n')
+
+        for i, nd in enumerate(self.nodes):
+            nd=nd[0].split()
+            ln = [str(i+1),str(nd[2]),str(nd[0]),str(nd[1])] # Making x-axis as beam axis
+            file.write(' '.join(ln) + '\n')
+
+        file.write('$EndNodes\n$Elements\n')
+
+    #    newNumEls = np.max(segment_element_labels)
+        file.write(str(self.num_elements) + '\n')
+
+        for j,eli in enumerate(self.elements):
+
+                ln = [str(j+1)]
+                ln.append('5')
+                ln.append('2')
+                ln.append(str(1))
+                ln.append(str(1))
+                ell=eli[0].split()
+                for n in ell:
+                    ln.append(n)
+                file.write(' '.join(ln) + '\n')
+
+        file.write('$EndElements\n')
+        file.close()
+
+        self._generate_layup_id()
+        self._generate_material_database()
+
+        # initialize segmentmesh object
+        segment_mesh = SolidBladeSegmentMesh(
+        #    segment_node_labels=segment_node_labels,
+         #   segment_element_labels=segment_element_labels,
+         #   segment_element_layer_id=segment_element_layer_id,
+            segment_index=segment_index,
+        #    elLayID=self.elLayID,
+            parent_blade_mesh=self,
+         #   mat_param=self.material_database
+            msh_file=filename)
+        return segment_mesh
+
+
+class ShellBladeSegmentMesh():
+    """A class representing a segment of a wind turbine blade mesh.
+
+    A segment is defined as the part of the blade between two fixed points along the blade span.
+    Given a set of N span points along the blade, there are N-1 segments defined between each consecutive pair
+    of span points. For example, the segment indexed by 0 is defined between the span points indexed by 0 and 1.
+
+    This class manages the data and methods for analyzing the structural properties of a blade segment,
+    including computing ABD matrices and stiffness properties.
+
+    Attributes
+    ----------
+    mesh : dolfinx.mesh.Mesh
+        The FEniCS/DOLFINx mesh object for this segment
+    subdomains : dolfinx.mesh.MeshTags
+        Tags identifying different regions/materials in the mesh
+    left_submesh : dict
+        Data for the left boundary submesh
+    right_submesh : dict
+        Data for the right boundary submesh
+    layup_database : dict
+        Database containing layup information including:
+            - mat_names: list of material names
+            - thick: list of layer thicknesses
+            - angle: list of fiber angles
+            - nlay: list of number of layers
+    """
+    def __init__(
+        self,
+        segment_node_labels,
+        segment_element_labels,
+        segment_element_layer_id,
+        segment_index,
+        layup_database,
+        parent_blade_mesh,
+        msh_file):
+        """Initialize a SegmentMesh object.
+
+        Parameters
+        ----------
+        segment_node_labels : array[int]
+            Labels for nodes in this segment
+        segment_element_labels : array[int]
+            Labels for elements in this segment
+        segment_element_layer_id : array[int]
+            Layer IDs for each element
+        segment_index : int
+            Index of this segment in the blade
+        layup_database : dict
+            Database containing layup information
+        parent_blade_mesh : BladeMesh
+            Reference to the parent blade mesh object
+        msh_file : str
+            Path to the GMSH mesh file
+        """
+        self.segment_node_labels = segment_node_labels
+        self.segment_element_labels = segment_element_labels
+        self.segment_element_layer_id = segment_element_layer_id
+        self.segment_index = segment_index
+        self.layup_database = layup_database
+        self.blade_mesh = parent_blade_mesh
+
+        self.mesh, self.subdomains, self.boundaries = gmshio.read_from_msh(msh_file, MPI.COMM_WORLD,0, gdim=3)
+        self.original_cell_index = self.mesh.topology.original_cell_index # Original cell Index from mesh file
+        lnn = self.subdomains.values[:]-1
+        self.num_cells = self.mesh.topology.index_map(self.mesh.topology.dim).size_local
+        cells = np.arange(self.num_cells, dtype=np.int32)
+
+        self.subdomains = dolfinx.mesh.meshtags(self.mesh, self.mesh.topology.dim, cells, np.array(lnn,dtype=np.int32))
+
+        self.tdim = self.mesh.topology.dim
+        self.fdim = self.tdim - 1
+
+        # self._generate_layup_data()
+        self._build_local_orientations()
+        self._build_boundary_submeshdata()
+
+        return
+
+
+    def _build_local_orientations(self):
+        """Build local orientation vectors for each element.
+
+        This method constructs the local coordinate system for each element
+        based on the element orientations provided in the mesh data.
+
+        Returns
+        -------
+        tuple
+            Three dolfinx.fem.Function objects representing the local coordinate
+            vectors (e1, e2, e3) for each element.
+        """
+        # Local Orientation (DG0 function) of quad mesh element (from yaml data)
+        VV = dolfinx.fem.functionspace(
+            self.mesh, basix.ufl.element(
+            "DG", self.mesh.topology.cell_name(),
+            0, shape=(3, )))
+        EE1, EE2, N = dolfinx.fem.Function(VV), dolfinx.fem.Function(VV), dolfinx.fem.Function(VV)
+
+        orientations = []
+        for i, eo in enumerate(self.blade_mesh.element_orientations):
+            if(self.segment_element_labels[i] > -1):
+                o = []
+                for k in range(9):
+                    o.append(eo[k])
+                orientations.append(o)
+
+        # Store orientation for each element
+        # TODO: clarify variable names. why N?
+        for k, ii in enumerate(self.original_cell_index):
+        # Storing data to DG0 functions 
+            EE2.x.array[3*k], EE2.x.array[3*k+1], EE2.x.array[3*k+2] = orientations[ii][5], orientations[ii][3], orientations[ii][4]  # e2
+            N.x.array[3*k], N.x.array[3*k+1], N.x.array[3*k+2] = orientations[ii][8], orientations[ii][6], orientations[ii][7]   #  e3 
+            EE1.x.array[3*k], EE1.x.array[3*k+1], EE1.x.array[3*k+2] = orientations[ii][2], orientations[ii][0], orientations[ii][1]  # e1    outward normal 
+        self.EE1 = EE1
+        self.N = N
+        self.EE2 = EE2
+        frame = [EE1,EE2,N]
+        self.frame = frame
+        return frame
+
+
+    def _build_boundary_submeshdata(self):
+        """Build submesh data for the left and right boundaries.
+
+        This method extracts and processes the mesh data for the boundary
+        regions of the segment, which are needed for applying boundary
+        conditions and computing boundary stiffness properties.
+
+        Returns
+        -------
+        tuple
+            Two dictionaries containing the submesh data for the left and
+            right boundaries respectively.
+        """
+        pp = self.mesh.geometry.x
+
+        is_left_boundary, is_right_boundary = opensg.utils.shell.generate_boundary_markers(
+            min(pp[:,0]), max(pp[:,0]))
+
+        left_facets = dolfinx.mesh.locate_entities_boundary(
+            self.mesh, dim=self.fdim, marker=is_left_boundary)
+        right_facets = dolfinx.mesh.locate_entities_boundary(
+            self.mesh, dim=self.fdim, marker=is_right_boundary)
+
+        left_mesh, left_entity_map, left_vertex_map, left_geom_map = dolfinx.mesh.create_submesh(
+            self.mesh, self.fdim, left_facets)
+        right_mesh, right_entity_map, right_vertex_map, right_geom_map = dolfinx.mesh.create_submesh(
+            self.mesh, self.fdim, right_facets)
+
+        self.left_submesh = {
+            "mesh": left_mesh,
+            "entity_map": left_entity_map,
+            "vertex_map": left_vertex_map,
+            "geom_map": left_geom_map,
+            "marker": is_left_boundary}
+
+        self.right_submesh = {
+            "mesh": right_mesh,
+            "entity_map": right_entity_map,
+            "vertex_map": right_vertex_map,
+            "geom_map": right_geom_map,
+            "marker": is_right_boundary}
+
+        self.mesh.topology.create_connectivity(2,1)  # (quad mesh topology, boundary(1D) mesh topology)
+        cell_of_facet_mesh = self.mesh.topology.connectivity(2,1)
+
+        # NOTE: found a different way to determine connectivity that didn't need this block -klb
+        # Cell to Edge connectivity
+        # cell_edge_map = []
+        # for i in range(self.num_cells):
+        #     c = []
+        #     for k in range(4): # 4 is used as number of edges in a quad element
+        #         c.append((cell_of_facet_mesh.array[4*i+k])) 
+        #     cell_edge_map.append(c)
+        # cell_edge_map = np.ndarray.flatten(np.array(cell_edge_map))
+
+        # generate subdomains
+        def _build_boundary_subdomains(boundary_meshdata):
+            boundary_mesh = boundary_meshdata["mesh"]
+            boundary_entity_map = boundary_meshdata["entity_map"]
+            boundary_marker = boundary_meshdata["marker"]
+            boundary_VV = dolfinx.fem.functionspace(
+                boundary_mesh, basix.ufl.element("DG", boundary_mesh.topology.cell_name(), 0, shape=(3, )))
+
+            boundary_e1 = dolfinx.fem.Function(boundary_VV)
+            boundary_e2 = dolfinx.fem.Function(boundary_VV)
+            boundary_n = dolfinx.fem.Function(boundary_VV)
+
+            boundary_facets = dolfinx.mesh.locate_entities(boundary_mesh, self.fdim, boundary_marker)
+
+            # TODO: review the subdomain assingments with akshat
+            boundary_subdomains = []
+            for i, xx in enumerate(boundary_entity_map):
+                # assign subdomain
+                # 4 is for number of nodes in quad element
+                # NOTE: we should find a different way to do this that doesn't assume quad elements if
+                #    we plan to expand to other elements. -klb
+                idx = int(np.where(cell_of_facet_mesh.array==xx)[0]/4)
+                boundary_subdomains.append(self.subdomains.values[idx])
+                # assign orientation
+                for j in range(3):
+                    boundary_e1.x.array[3*i+j] = self.EE1.x.array[3*idx+j]
+                    boundary_e2.x.array[3*i+j] = self.EE2.x.array[3*idx+j]
+                    boundary_n.x.array[3*i+j] = self.N.x.array[3*idx+j]
+
+            boundary_frame = [boundary_e1, boundary_e2, boundary_n]
+            boundary_subdomains = np.array(boundary_subdomains, dtype=np.int32)
+            boundary_num_cells = boundary_mesh.topology.index_map(boundary_mesh.topology.dim).size_local
+            boundary_cells = np.arange(boundary_num_cells, dtype=np.int32)
+            boundary_subdomains = dolfinx.mesh.meshtags(boundary_mesh, boundary_mesh.topology.dim, boundary_cells, boundary_subdomains)
+
+            return boundary_subdomains, boundary_frame, boundary_facets
+            # Mapping the orinetation data from quad mesh to boundary. The alternative is to use local_frame_1D(self.left_submesh["mesh"]).
+            # Either of both can be used in local_boun subroutine 
+
+        self.left_submesh["subdomains"], self.left_submesh["frame"], self.left_submesh["facets"] = _build_boundary_subdomains(self.left_submesh)
+        self.right_submesh["subdomains"], self.right_submesh["frame"], self.right_submesh["facets"] = _build_boundary_subdomains(self.right_submesh)
+
+        return self.left_submesh, self.right_submesh
+
+    def compute_ABD(self):
+        """Compute the ABD (stiffness) matrices for the segment.
+
+        This method computes the ABD matrices that relate forces and moments
+        to strains and curvatures for each unique layup in the segment.
+
+        Returns
+        -------
+        list
+            List of 6x6 numpy arrays representing the ABD matrices for each
+            unique layup in the segment.
+        """
+        nphases = max(self.subdomains.values[:]) + 1
+        ABD_ = []
+        for i in range(nphases):
+            ABD_.append(opensg.compute_ABD_matrix(
+                thick=self.layup_database["thick"][i],
+                nlay=self.layup_database["nlay"][i],
+                mat_names=self.layup_database["mat_names"][i],
+                angle=self.layup_database["angle"][i],
+                material_database=self.blade_mesh.material_database
+                ))
+
+        print('Computed',nphases,'ABD matrix')
+        return ABD_
+
+
+    # def plot(self):
+    #     import pyvista
+    #     pyvista.start_xvfb()
+    #     u_topology, u_cell_types, u_geometry=dolfinx.plot.vtk_mesh(self.blade_mesh.mesh,self. mesh.blade_mesh.topology.dim)
+    #     grid = pyvista.UnstructuredGrid(u_topology, u_cell_types, u_geometry)
+    #     grid.cell_data["Marker"] = self.blade_mesh.subdomains.values[:]
+    #     grid.set_active_scalars("Marker")
+    #     u_plotter = pyvista.Plotter()
+    #     u_plotter.add_mesh(grid)
+    #     u_plotter.show_axes()
+    #     #u_plotter.view_xy() # z is beam axis
+    #     u_plotter.show()
+
     def compute_stiffness_EB(self, ABD):
+        """Compute the Euler-Bernoulli beam stiffness matrix.
+
+        Parameters
+        ----------
+        ABD : list
+            List of ABD matrices for each unique layup
+
+        Returns
+        -------
+        numpy.ndarray
+            4x4 stiffness matrix for Euler-Bernoulli beam theory
+        """
         # extract object data
         mesh = self.mesh
         frame = self.frame
         subdomains = self.subdomains
-        tdim=mesh.topology.dim
-        fdim = tdim - 1
-        nphases = max(self.subdomains.values[:]) + 1
-        
-        pp = mesh.geometry.x # point data
-        x_min, x_max=min(pp[:,0]), max(pp[:,0])
-        # Initialize terms
-        e_l, V_l, dvl, v_l, x_l, dx_l = opensg.local_boun(
-            self.left_submesh["mesh"], self.left_submesh["frame"] ,self.left_submesh["subdomains"])
-        
-        e_r, V_r, dvr, v_r, x_r, dx_r = opensg.local_boun(
-            self.right_submesh["mesh"], self.right_submesh["frame"] ,self.right_submesh["subdomains"])
-        
-        self.left_submesh["nullspace"] = opensg.compute_nullspace(V_l)
-        self.right_submesh["nullspace"] = opensg.compute_nullspace(V_r)
-        
-        A_l = opensg.A_mat(ABD, e_l,x_l,dx_l,self.left_submesh["nullspace"],v_l,dvl, nphases)
-        A_r = opensg.A_mat(ABD, e_r,x_r,dx_r,self.right_submesh["nullspace"],v_r,dvr, nphases)
-        
-        V0_l = opensg.solve_boun(ABD, self.left_submesh, nphases)
-        
-        V0_r = opensg.solve_boun(ABD, self.right_submesh, nphases)
-        # The local_frame_l(self.left_submesh["mesh"]) can be replaced with frame_l, if we want to use mapped orientation from given direction cosine matrix (orien mesh data-yaml)
 
-        # Quad mesh
-        e, V, dv, v_, x, dx = opensg.local_boun(mesh, frame, subdomains)
-        V0, Dle, Dhe, Dhd, Dld, D_ed, D_dd, D_ee, V1s = opensg.initialize_array(V)
-        mesh.topology.create_connectivity(1, 2)
-        self.left_submesh["mesh"].topology.create_connectivity(1, 1)
-        self.right_submesh["mesh"].topology.create_connectivity(1, 1)
+        D_eff = opensg.compute_stiffness_EB_blade_segment(
+            ABD,
+            mesh,
+            frame,
+            subdomains,
+            self.left_submesh,
+            self.right_submesh)
 
-        # Obtaining coefficient matrix AA and BB with and without bc applied.
-        # Note: bc is applied at boundary dofs. We define v2a containing all dofs of entire wind blade.
-        boundary_dofs = locate_dofs_topological(V, fdim, np.concatenate((self.right_submesh["entity_map"], self.left_submesh["entity_map"]), axis=0))
-        F2=sum([dot(dot(as_tensor(ABD[i]), opensg.gamma_h(e,x,dv)), opensg.gamma_h(e,x,v_))*dx(i) for i in range(nphases)])  
-        v2a=Function(V) # b default, v2a has zero value for all. 
-        bc = dolfinx.fem.dirichletbc(v2a, boundary_dofs) # This shows only boundary_dofs are taken for v2a under bc, which are zero (known) as input.
-        a = form(F2)
-        
-        B = assemble_matrix(a)   # Obtain coefficient matrix without BC applied: BB
-        B.assemble()
-        ai, aj, av=B.getValuesCSR()
-        BB = csr_matrix((av, aj, ai))
-        BB = BB.toarray()  
-        
-        A = assemble_matrix(a,[bc])  # Obtain coefficient matrix with BC applied: AA
-        A.assemble()
-        ai, aj, av=A.getValuesCSR()
-        AA = csr_matrix((av, aj, ai))
-        AA=AA.toarray()
-        avg=np.trace(AA)/len(AA)     
-
-        # averaging is done so that all terms are of same order. Note after appliying bc at [A=assemble_matrix(a,[bc])], the dofs of
-        # coefficientmatrix has only 1 replaced at that dofs. 
-        for i,xx in enumerate(av):
-            if xx==1:
-                av[i]=avg        
-                                
-        AA_csr = csr_matrix((av, aj, ai))
-        AAA = AA_csr.toarray() 
-        AA = scipy.sparse.csr_matrix(AAA) 
-
-        # Assembly
-        # Running for 4 different F vector. However, F has bc applied to it where, stored known values of v2a is provided for each loop (from boun solve).
-        for p in range(4): # 4 load cases meaning 
-            # Boundary 
-            v2a = Function(V)
-            v2a = opensg.dof_mapping_quad(V, v2a,V_l,V0_l[:,p], self.left_submesh["facets"], self.left_submesh["entity_map"]) 
-            v2a = opensg.dof_mapping_quad(V, v2a,V_r,V0_r[:,p], self.right_submesh["facets"], self.right_submesh["entity_map"])  
-            
-            # quad mesh
-            F2=sum([dot(dot(as_tensor(ABD[i]),opensg.gamma_e(e,x)[:,p]), opensg.gamma_h(e,x,v_))*dx(i) for i in range(nphases)])  
-            bc = dolfinx.fem.dirichletbc(v2a, boundary_dofs)
-            F = petsc.assemble_vector(form(rhs(F2)))
-            apply_lifting(F, [a], [[bc]]) # apply bc to rhs vector (Dhe)
-            set_bc(F, [bc])
-            with F.localForm() as local_F:
-                for i in boundary_dofs:
-                    for k in range(3):
-                        # F[3*i+k]=avg*F[3*i+k] # normalize small terms
-                        local_index = 3 * i + k
-                        local_F[local_index] = avg * local_F[local_index]
-                    
-            V0[:,p]=  scipy.sparse.linalg.spsolve(AA, F, permc_spec=None, use_umfpack=True) # obtain sol: E* V1s = b*
-            Dhe[:,p]= scipy.sparse.csr_array(BB).dot(V0[:,p])
-            
-        D1 = np.matmul(V0.T,-Dhe) 
-        for s in range(4):
-            for k in range(4): 
-                f = dolfinx.fem.form(sum([dot(dot(opensg.gamma_e(e,x).T,as_tensor(ABD[i])),opensg.gamma_e(e,x))[s,k]*dx(i) for i in range(nphases)]))
-                D_ee[s,k]=dolfinx.fem.assemble_scalar(f)
-        L = (x_max - x_min)
-        D_eff= D_ee + D1
-        D_eff=D_eff/L  # L is divided because of 3D shell mesh and corresponding beam length need to divided.
-        #--------------------------------------Printing Output Data---------------------------------------
-        print('  ')  
-        print('Stiffness Matrix')
-        np.set_printoptions(precision=4)
-        print(np.around(D_eff))
-        
         return D_eff
-"""
-        
 
-class MeshData:
-    def __init__(self, mesh):
-        # What data is required to start?
-        
-        pass
-    
-    def add_submesh_attributes(self,
-        entity_map,
-        vertex_map,
-        geom_map,
-        submesh_facets):
-        pass
-    
-    def process_data(self):
-        pass
-        # what downstream data can be generated with the starting data?
-        
-    def process_data_with_args(self, args):
-        pass
-        # what data needs external information to be generated?
-        # how do we track that this data is not soley derived from the init data?
-        
-            # Mapping the orinetation data from quad mesh to boundary. The alternative is to use local_frame_1D(self.left_submesh["mesh"]).
-            # Either of both can be used in local_boun subroutine 
-        
-            # self.left_submesh["subdomains"], self.left_submesh["frame"], self.left_submesh["facets"] = subdomains_boundary(
-            #     self.left_submesh["mesh"], is_left_boundary, self.left_submesh["entity_map"]) 
-            # self.right_submesh["subdomains"], self.right_submesh["frame"], self.right_submesh["facets"] = subdomains_boundary(
-            #     self.right_submesh["mesh"], is_right_boundary, self.right_submesh["entity_map"])
-            
-            
-def _generate_submesh_subdomains(submesh_data, mesh_data):
-        mesh_data.mesh.topology.create_connectivity(2,1)  # (quad mesh topology, boundary(1D) mesh topology)
-        cell_of_facet_mesh = mesh_data.mesh.topology.connectivity(2,1)
-        submesh_VV = dolfinx.fem.functionspace(
-            submesh_data.mesh, basix.ufl.element("DG", submesh_data.mesh.topology.cell_name(), 0, shape=(3, )))
-        
-        submesh_e1 = dolfinx.fem.Function(submesh_VV)
-        submesh_e2 = dolfinx.fem.Function(submesh_VV)
-        submesh_n = dolfinx.fem.Function(submesh_VV)
-        
-        submesh_facets = dolfinx.mesh.locate_entities(submesh_data.mesh, self.fdim, submesh_marker)
-        submesh_subdomains = []
-        for i, xx in enumerate(submesh_data.entity_map):
-            # assign subdomain
-            idx = int(np.where(cell_of_facet_mesh.array==xx)[0]/4) # 4 is for number of nodes in quad element
-            submesh_subdomains.append(mesh_data.subdomains.values[idx])
-            # assign orientation
-            for j in range(3):
-                submesh_e1.x.array[3*i+j] = mesh_data.EE1.x.array[3*idx+j]
-                submesh_e2.x.array[3*i+j] = mesh_data.EE2.x.array[3*idx+j]
-                submesh_n.x.array[3*i+j] = mesh_data.N.x.array[3*idx+j]
+    def compute_boundary_stiffness_timo(self, ABD):
+        """Compute the Timoshenko beam stiffness matrices for the boundaries.
 
-        submesh_frame = [submesh_e1, submesh_e2, submesh_n]
-        submesh_subdomains = np.array(submesh_subdomains, dtype=np.int32)
-        submesh_num_cells = submesh_data.mesh.topology.index_map(submesh_data.mesh.topology.dim).size_local 
-        submesh_cells = np.arange(submesh_num_cells, dtype=np.int32)
-        submesh_subdomains = dolfinx.mesh.meshtags(
-            submesh_data.mesh, submesh_data.mesh.topology.dim, submesh_cells, submesh_subdomains)
-        
-        return submesh_subdomains, submesh_frame, submesh_facets
+        Parameters
+        ----------
+        ABD : list
+            List of ABD matrices for each unique layup
+
+        Returns
+        -------
+        tuple
+            Left and right boundary 6x6 Timoshenko stiffness matrices
+        """
+        left_stiffness = core.compute_timo_boun(ABD, self.left_submesh)[1]
+
+        right_stiffness = core.compute_timo_boun(
+            ABD,
+            self.right_submesh["mesh"],
+            self.right_submesh["subdomains"],
+            self.right_submesh["frame"],
+            self.nullspace, # quad nullspace
+            self.right_submesh["nullspace"],
+            self.nphases)[1]
+
+        return left_stiffness, right_stiffness
+
+    def compute_stiffness(self, ABD):
+        """Compute all stiffness matrices for the segment.
+
+        This method computes both the Euler-Bernoulli and Timoshenko
+        stiffness matrices for the segment and its boundaries.
+
+        Parameters
+        ----------
+        ABD : list
+            List of ABD matrices for each unique layup
+
+        Returns
+        -------
+        tuple
+            Contains:
+            - segment_timo_stiffness: 6x6 Timoshenko stiffness matrix
+            - segment_eb_stiffness: 4x4 Euler-Bernoulli stiffness matrix
+            - l_timo_stiffness: 6x6 left boundary Timoshenko stiffness matrix
+            - r_timo_stiffness: 6x6 right boundary Timoshenko stiffness matrix
+        """
+        return core.compute_stiffness(
+            ABD=ABD,
+            mesh=self.mesh,
+            subdomains=self.subdomains,
+            l_submesh=self.left_submesh,
+            r_submesh=self.right_submesh
+        )
+
